requirements.txt
pydantic
uvicorn
fastapi
passlib[bcrypt]
python-jose[cryptography]
starlette

